refactor: File foldering, TableView의 delegate,datasource를 MainViewController로 이동
